error handling

+++ app/src/main/java/com/example/dogapiapp/data/repository/DogBreedRepositoryImpl.kt
@@ -3,6 +3,8 @@ package com.example.dogapiapp.data.repository
 import androidx.paging.PagingData
 import com.example.dogapiapp.data.local.dao.DogBreedDao
 import com.example.dogapiapp.data.local.model.DogBreedDbModel
+import com.example.dogapiapp.data.local.toDogBreedDbModelList
+import com.example.dogapiapp.data.remote.ApiResult
 import com.example.dogapiapp.data.remote.remotedatasource.DogBreedsRemoteDataSource
 import io.reactivex.rxjava3.core.Flowable
 import io.reactivex.rxjava3.core.Single
@@ -11,7 +13,7 @@ import javax.inject.Inject
 class DogBreedRepositoryImpl @Inject constructor(
     private val dogBreedDao: DogBreedDao,
     private val dataSource: DogBreedsRemoteDataSource,
-): DogBreedRepository {
+) : DogBreedRepository {
 
     override fun getDogBreedsWithPagination(): Flowable<PagingData<DogBreedDbModel>> {
         return dataSource.getDogBreedsWithPagination()
@@ -21,20 +23,37 @@ class DogBreedRepositoryImpl @Inject constructor(
         dogBreedDao.insertAll(dogBreeds)
     }
 
-    override fun getDogBreedById(id: Int): Single<DogBreedDbModel> {
+    override fun getDogBreedById(id: Int): Single<RepoResult<DogBreedDbModel>> {
         return dogBreedDao.getDogBreedById(id)
+            .map<RepoResult<DogBreedDbModel>> {
+                RepoResult.Saved(it)
+            }
+            .onErrorReturn {
+                RepoResult.Error(it.message.toString())
+            }
     }
 
     override fun getAllDogBreedsWithoutPaginationFromDb(): List<DogBreedDbModel> {
         return dogBreedDao.getDogBreedsFromDbWithoutPaging()
     }
 
-    override fun getAllDogBreedsWithoutPagination(): Single<List<DogBreedDbModel>> {
+    override fun getAllDogBreedsWithoutPagination(): Single<RepoResult<List<DogBreedDbModel>>> {
         return dataSource.getDogBreedsWithoutPagination()
-            .map {
-                dogBreedDao.deleteAllDogBreeds()
-                dogBreedDao.insertAll(it)
-                it
+            .map { apiResult ->
+                when (apiResult) {
+                    is ApiResult.Success -> {
+                        val dogBreeds = apiResult.data.toDogBreedDbModelList()
+                        dogBreedDao.deleteAllDogBreeds()
+                        dogBreedDao.insertAll(dogBreeds)
+                        RepoResult.Success(dogBreeds)
+                    }
+                    is ApiResult.Error -> {
+                        RepoResult.Error(apiResult.errorMessage)
+                    }
+                }
+            }
+            .onErrorReturn {
+                RepoResult.Error(it.message.toString())
             }
     }
